Major Rework
Major rework implementing all the changes I've made on the local version over the last few months. Primarily this is comprised of reworking the mil groups into the strategy/tactics/navy mutually exclusive options, swapping around a lot of admin stuff to be more balanced/worth taking, and the introduction of a potential privateering choice.